requirements.txt
# Core RAG stack
langchain>=0.2.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.5.0

# Embeddings (choose one or more depending on provider)
openai>=1.0.0
tiktoken>=0.7.0
sentence-transformers>=2.2.2

# Vector math + performance
numpy>=1.24.0
scipy>=1.11.0

# API serving
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Environment + config
python-dotenv>=1.0.0
pydantic>=2.7.0


groq>=0.9.0
langchain-groq>=0.1.0